Nano-optimize: inline everything possible, use call vs callvirt where possible, remove assertions in Release mode (DirectBuffer in Spreads.Core).

+++ src/Spreads.IPC/LogBuffers.cs
@@ -25,7 +25,7 @@ namespace Spreads.IPC
             try
             {
                 long logLength = LogBufferDescriptor.PARTITION_COUNT *
-                                 (LogBufferDescriptor.TERM_META_DATA_LENGTH + termLength) +
+                                 (LogBufferDescriptor.TERM_META_DATA_LENGTH + (long)termLength) +
                                  LogBufferDescriptor.LOG_META_DATA_LENGTH;
                 termLength = LogBufferDescriptor.ComputeTermLength(logLength);
                 LogBufferDescriptor.CheckTermLength(termLength);
